Create DatePurchase
This is the second function of our model, which takes the list of utilities for each agent as input (which is generated through the 'TrueUtility' function, and outputs the day each agent will end up purchasing their Christmas shopping. 

I have purposely left this uncommented, as if I add in comments to the code then it would cease to work if/when copied to a local machine running Mathematica.